feat(cdk): add `uniqBy` utility method
