+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -21,7 +21,7 @@ What kind of change does this PR introduce?
 
 <!-- Please describe the current behavior that you are modifying, or link to a relevant issue. -->
 
-Issue Number: N/A
+Closes #
 
 ## What is the new behavior?
