+++ projects/cdk/utils/miscellaneous/index.ts
@@ -6,3 +6,4 @@ export * from './identity';
 export * from './is-present';
 export * from './mark-control-as-touched-and-validate';
 export * from './nullable-same';
+export * from './uniq-by';
